mysite2 - delete 추가

+++ src/main/java/kr/co/itcen/mysite/action/board/DeleteAction.java
@@ -18,44 +18,46 @@ public class DeleteAction implements Action {
 	@Override
 	public void execute(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		Long uNo = Long.parseLong(request.getParameter("uno")); // 게시글 만든 사용자 번호
-		
+		String text = request.getParameter("text");
+		System.out.println("유저번호: " + uNo);
+
 		HttpSession session = request.getSession();
 		if (session == null) {
 			WebUtils.redirect(request, response, request.getContextPath() + "/board?a=list");
 			return;
 		}
-		
+
 		UserVo authUser = (UserVo) session.getAttribute("authUser"); // 로그인한 사용자 정보
-		
+
 		if (authUser == null) {
 			WebUtils.redirect(request, response, request.getContextPath() + "/board?a=list");
 			return;
 		}
 		System.out.println("1 : " + authUser.getNo());
 		System.out.println("2 : " + uNo);
-		
+
 		if (authUser.getNo() != uNo) {
 			WebUtils.redirect(request, response, request.getContextPath() + "/board?a=list");
 			return;
 		}
-		
+
 		Long no = Long.parseLong(request.getParameter("no"));
-		String password = (String)session.getAttribute(authUser.getPassword());
-		
+		System.out.println("게시글번호: " + no);
 		BoardVo boardVo = new BoardVo();
 		boardVo.setNo(no);
-		boardVo.setPassword(password);
+		boardVo.setuNo(uNo);
 //		System.out.println(boardVo.getContents());
 //		System.out.println(boardVo.getTitle());
 //		System.out.println(boardVo.getuNo());
 //
-		new BoardDao().delete(boardVo);
+		if ("삭제하기".equals(text)) {
+			new BoardDao().delete(boardVo);
+		}
+		
 		System.out.println("찍히냐");
 //
 		WebUtils.redirect(request, response, request.getContextPath() + "/board?a=list");
-		
-		
-		
+
 	}
 
 }
